==> 03_main_hypothesis_jasp_code.R <==
# JASP CODE
# 03_main_hypothesis

# Bayesian borrelation
jaspRegression::CorrelationBayesian(
  data = NULL,
  version = "0.17.2",
  alternative = "twoSided",
  bayesFactorReport = TRUE,
  bayesFactorType = "BF10",
  bfRobustnessPlot = FALSE,
  bfRobustnessPlotAdditionalInfo = TRUE,
  bfSequentialPlot = FALSE,
  bfSequentialPlotAdditionalInfo = TRUE,
  ci = FALSE,
  ciLevel = 0.95,
  kendall = TRUE,
  matrixPlot = FALSE,
  matrixPlotDensity = FALSE,
  matrixPlotPosterior = FALSE,
  naAction = "pairwise",
  pairsMethod = "pearson",
  pairwiseDisplay = FALSE,
  pearson = FALSE,
  plotHeight = 320,
  plotWidth = 480,
  posteriorMedian = FALSE,
  priorPosteriorPlot = FALSE,
  priorPosteriorPlotAdditionalEstimationInfo = TRUE,
  priorPosteriorPlotAdditionalTestingInfo = TRUE,
  priorWidth = "1.0",
  sampleSize = FALSE,
  scatterPlot = TRUE,
  scatterPlotAdditionalInfo = FALSE,
  seed = 1,
  setSeed = FALSE,
  spearman = FALSE,
  supportCorrelationFlagged = TRUE,
  variablePairs = list(),
  variables = list("ace_postimp_ord_4", "phq9_postimp", "gad7_postimp", "bilat_hippo", "bilat_amy", "dlpfc", "age", "ses_index", "sex", "tiv"))

# Bayesian Ancova PHQ-9
jaspAnova::AncovaBayesian(
  data = NULL,
  version = "0.17.2",
  formula = phq9_postimp ~ age + sex + ace_postimp_ord_4,
  isNuisance = ~ age + sex,
  covariates = list("age", "sex"),
  bayesFactorOrder = "nullModelTop",
  bayesFactorType = "BF10",
  bernoulliParameter = 0.5,
  betaBinomialParameterA = 1,
  betaBinomialParameterB = 1,
  castilloParameterU = 1,
  cauchyPriorScaleCovariates = 0.354,
  cauchyPriorScaleFixedEffects = 0.5,
  cauchyPriorScaleRandomEffects = 1,
  credibleInterval = 0.95,
  criTable = TRUE,
  customPriorSpecification = list(list(components = "ace_postimp_ord_4", inclusionProbability = 0.5, scaleFixedEffects = 0.5)),
  descriptivePlotCi = TRUE,
  descriptivePlotCiLevel = 0.95,
  descriptivePlotHorizontalAxis = "ace_postimp_ord_4",
  descriptivePlotSeparateLines = "",
  descriptivePlotSeparatePlot = "",
  descriptives = TRUE,
  effects = FALSE,
  effectsType = "allModels",
  enforcePrincipleOfMarginalityFixedEffects = TRUE,
  enforcePrincipleOfMarginalityRandomSlopes = FALSE,
  groupPosterior = "grouped",
  hideNuisanceParameters = TRUE,
  integrationMethod = "automatic",
  legacyResults = FALSE,
  modelAveragedPosteriorPlot = FALSE,
  modelPrior = "uniform",
  modelsShown = "unlimited",
  numModelsShown = 10,
  plotHeight = 320,
  plotWidth = 480,
  postHocNullControl = FALSE,
  postHocTerms = "ace_postimp_ord_4",
  posteriorEstimates = TRUE,
  priorSpecificationMode = "acrossParameters",
  qqPlot = TRUE,
  rainCloudHorizontalAxis = "",
  rainCloudHorizontalDisplay = FALSE,
  rainCloudSeparatePlots = "",
  rainCloudYAxisLabel = "",
  rsqPlot = FALSE,
  samplesMCMC = 1000,
  samplesNumericAccuracy = 10000,
  samplingMethodMCMC = "auto",
  samplingMethodNumericAccuracy = "auto",
  seed = 1,
  setSeed = FALSE,
  singleModelCriTable = TRUE,
  singleModelEstimates = TRUE,
  singleModelGroupPosterior = "grouped",
  singleModelPosteriorPlot = FALSE,
  singleModelQqPlot = TRUE,
  singleModelRsqPlot = FALSE,
  singleModelTerms = ~ age + sex + ace_postimp_ord_4,
  wilsonParameterLambda = 1)

# Bayesian Ancova GAD-7
jaspAnova::AncovaBayesian(
  data = NULL,
  version = "0.17.2",
  formula = gad7_postimp ~ sex + age + ace_postimp_ord_4,
  isNuisance = ~ sex + age,
  covariates = list("age", "sex"),
  bayesFactorOrder = "nullModelTop",
  bayesFactorType = "BF10",
  bernoulliParameter = 0.5,
  betaBinomialParameterA = 1,
  betaBinomialParameterB = 1,
  castilloParameterU = 1,
  cauchyPriorScaleCovariates = 0.354,
  cauchyPriorScaleFixedEffects = 0.5,
  cauchyPriorScaleRandomEffects = 1,
  credibleInterval = 0.95,
  criTable = TRUE,
  customPriorSpecification = list(list(components = "ace_postimp_ord_4", inclusionProbability = 0.5, scaleFixedEffects = 0.5)),
  descriptivePlotCi = TRUE,
  descriptivePlotCiLevel = 0.95,
  descriptivePlotHorizontalAxis = "ace_postimp_ord_4",
  descriptivePlotSeparateLines = "",
  descriptivePlotSeparatePlot = "",
  descriptives = TRUE,
  effects = FALSE,
  effectsType = "allModels",
  enforcePrincipleOfMarginalityFixedEffects = TRUE,
  enforcePrincipleOfMarginalityRandomSlopes = FALSE,
  groupPosterior = "grouped",
  hideNuisanceParameters = TRUE,
  integrationMethod = "automatic",
  legacyResults = FALSE,
  modelAveragedPosteriorPlot = FALSE,
  modelPrior = "uniform",
  modelsShown = "unlimited",
  numModelsShown = 10,
  plotHeight = 320,
  plotWidth = 480,
  postHocNullControl = FALSE,
  postHocTerms = "ace_postimp_ord_4",
  posteriorEstimates = TRUE,
  priorSpecificationMode = "acrossParameters",
  qqPlot = TRUE,
  rainCloudHorizontalAxis = "",
  rainCloudHorizontalDisplay = FALSE,
  rainCloudSeparatePlots = "",
  rainCloudYAxisLabel = "",
  rsqPlot = FALSE,
  samplesMCMC = 1000,
  samplesNumericAccuracy = 10000,
  samplingMethodMCMC = "auto",
  samplingMethodNumericAccuracy = "auto",
  seed = 1,
  setSeed = FALSE,
  singleModelCriTable = TRUE,
  singleModelEstimates = TRUE,
  singleModelGroupPosterior = "grouped",
  singleModelPosteriorPlot = FALSE,
  singleModelQqPlot = TRUE,
  singleModelRsqPlot = FALSE,
  singleModelTerms = ~ sex + age + ace_postimp_ord_4,
  wilsonParameterLambda = 1)
